--- lib/client/mongodb/IBaseRepository.js
@@ -1,3 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=IBaseRepository.js.map
