+++ lib/index.js
@@ -26,7 +26,6 @@ __exportStar(require("./encryption/AesEncryption"), exports);
 __exportStar(require("./services/secret/SecretManager"), exports);
 __exportStar(require("./services/s3/S3Helper"), exports);
 __exportStar(require("./services/sqs/SqsHelper"), exports);
-__exportStar(require("./services/authorization/JWTBaseAuthorization"), exports);
 __exportStar(require("./services/authorization/JWTAuthorization"), exports);
 __exportStar(require("./services/authorization/WebhookAuthorization"), exports);
 __exportStar(require("./services/lambda/authenticator/jwt/JwtAuthenticator"), exports);
