+++ src/index.ts
@@ -14,7 +14,6 @@ export * from './services/secret/SecretManager';
 export * from './services/s3/S3Helper';
 export * from './services/sqs/SqsHelper';
 
-export * from './services/authorization/JWTBaseAuthorization';
 export * from './services/authorization/JWTAuthorization';
 export * from './services/authorization/WebhookAuthorization';
